ultimoscambiosbd

+++ Laragon/www/lelsystem/LELSYSTEM-V2/database/migrations/2022_08_28_032128_create_semestres_table.php
@@ -17,11 +17,11 @@ class CreateSemestresTable extends Migration
             $table->id('id_semestre')->unique();
             $table->double('nota');
             $table->boolean('estado');
-            $table->bigInteger('id_matricula');
+            $table->unsignedInteger('id_matricula');
             $table->foreign('id_matricula')->references('id_matricula')->on('matriculas');
-            $table->bigInteger('id_archivo_parcial1');
+            $table->unsignedInteger('id_archivo_parcial1');
             $table->foreign('id_archivo_parcial1')->references('id_archivo_parcial1')->on('archivos_parcial1');
-            $table->bigInteger('id_archivo_parcial2');
+            $table->unsignedInteger('id_archivo_parcial2');
             $table->foreign('id_archivo_parcial2')->references('id_archivo_parcial2')->on('archivos_parcial2');
             $table->timestamps();
